requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
requests==2.31.0
python-dotenv==1.0.0
supabase==1.0.3
twilio==8.9.0

# ML Dependencies - UPDATED VERSIONS
tensorflow==2.15.0  # Updated from 2.13.0 to 2.15.0
librosa==0.10.1
numpy==1.24.3
scikit-learn==1.3.0
scipy==1.11.1
pandas==2.0.3
soundfile==0.12.1